+++ src/main/java/de/maxhenkel/corpse/integration/waila/PluginCorpse.java
@@ -1,5 +1,5 @@
 package de.maxhenkel.corpse.integration.waila;
-/*
+
 import de.maxhenkel.corpse.entities.CorpseEntity;
 import mcp.mobius.waila.api.IRegistrar;
 import mcp.mobius.waila.api.IWailaPlugin;
@@ -18,4 +18,3 @@ public class PluginCorpse implements IWailaPlugin {
     }
 
 }
-*/